feat: set base path
