+++ src/pages/404.tsx
@@ -1,13 +1,13 @@
-import * as React from "react"
-import Layout from "../components/layout"
+import * as React from 'react';
+import Layout from '../components/layout';
 
 // markup
 const NotFoundPage = () => {
-  return (
-    <Layout pageTitle="Page not found">
-      <h1>404! PAGE NOT FOUND!</h1>
-    </Layout>
-  )
-}
+    return (
+        <Layout pageTitle="Page not found">
+            <h1>404! PAGE NOT FOUND!</h1>
+        </Layout>
+    );
+};
 
-export default NotFoundPage
+export default NotFoundPage;
